Removed improper check
The check would fail for columns that contain `(500c)` in the column name. The check will need to be disabled until I figure how to correctly parse for column numbers, vs parts of the column name. One option would be to avoid the `XASSERT` when `ctrees_colnum` exceeds the total number of available columns.